+++ examples/basic/www/js/index.js
@@ -5,7 +5,7 @@ const app = {
     document.addEventListener(
       'deviceready',
       this.onDeviceReady.bind(this),
-      false
+      false,
     )
   },
 
@@ -13,10 +13,29 @@ const app = {
     this.receivedEvent('deviceready')
 
     this.showAds()
+    this.checkIsLoaded()
+  },
+
+  checkIsLoaded() {
+    return admob.interstitial.isLoaded().then(result => {
+      console.log(result)
+      return result
+    })
   },
 
   showAds() {
-    admob.banner.show({ id: 'test' }).catch(console.log)
+    const bannerTop = new admob.BannerAd({
+      adUnitId: 'ca-app-pub-3940256099942544/6300978111',
+    })
+    bannerTop.show({ position: 'top' })
+    const banner = new admob.BannerAd({
+      adUnitId: 'ca-app-pub-3940256099942544/2934735716',
+    })
+    banner.show({ position: 'bottom' })
+  },
+
+  showAdsOld() {
+    admob.banner.show({ id: 'test', npa: '1' }).catch(console.log)
     admob.interstitial
       .load({ id: 'test' })
       .then(() => admob.interstitial.show())