+++ packages/cordova/src/ios/AMSBanner.swift
@@ -2,9 +2,14 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
     var bannerView: GADBannerView!
     var adSize: GADAdSize!
     var position: String!
+    var constraintsToHide: [NSLayoutConstraint]!
 
-    var view: UIView {
+    var rootView: UIView {
         return self.plugin.viewController.view
+    }
+
+    var mainView: UIView {
+        return self.plugin.webView
     }
 
     init(id: Int, adUnitID: String, adSize: GADAdSize, position: String) {
@@ -12,6 +17,10 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
 
         self.adSize = adSize
         self.position = position
+        self.constraintsToHide = [
+            mainView.topAnchor.constraint(equalTo: rootView.topAnchor),
+            mainView.bottomAnchor.constraint(equalTo: rootView.bottomAnchor)
+        ]
     }
 
     deinit {
@@ -19,6 +28,7 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
     }
 
     func show(request: GADRequest) {
+        NSLayoutConstraint.deactivate(self.constraintsToHide)
         if bannerView != nil {
             bannerView.isHidden = false
         } else {
@@ -38,25 +48,26 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
             bannerView.rootViewController = nil
             bannerView.removeFromSuperview()
             bannerView = nil
+            NSLayoutConstraint.activate(self.constraintsToHide)
         }
     }
 
     func addBannerViewToView(_ bannerView: UIView) {
-        self.plugin.webView.translatesAutoresizingMaskIntoConstraints = false
+        mainView.translatesAutoresizingMaskIntoConstraints = false
         bannerView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(bannerView)
+        rootView.addSubview(bannerView)
         if #available(iOS 11.0, *) {
             positionBannerInSafeArea(bannerView)
 
             let background = UIView()
             background.translatesAutoresizingMaskIntoConstraints = false
             background.backgroundColor = .black
-            view.addSubview(background)
-            view.sendSubview(toBack: background)
+            rootView.addSubview(background)
+            rootView.sendSubviewToBack(background)
             NSLayoutConstraint.activate([
-                background.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-                background.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-                background.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+                background.leadingAnchor.constraint(equalTo: rootView.leadingAnchor),
+                background.trailingAnchor.constraint(equalTo: rootView.trailingAnchor),
+                background.bottomAnchor.constraint(equalTo: rootView.bottomAnchor),
                 background.topAnchor.constraint(equalTo: bannerView.topAnchor)
             ])
         } else {
@@ -66,49 +77,53 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
 
     @available (iOS 11, *)
     func positionBannerInSafeArea(_ bannerView: UIView) {
-        let guide: UILayoutGuide = view.safeAreaLayoutGuide
+        let guide: UILayoutGuide = rootView.safeAreaLayoutGuide
         var constraints = [
             bannerView.centerXAnchor.constraint(equalTo: guide.centerXAnchor),
-            self.plugin.webView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            self.plugin.webView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            bannerView.heightAnchor.constraint(equalToConstant: bannerView.frame.height),
+            mainView.widthAnchor.constraint(equalTo: rootView.widthAnchor)
         ]
-        if position == "top" {
+        if position == AMSBannerPosition.top {
+            let c = mainView.bottomAnchor.constraint(equalTo: rootView.bottomAnchor)
+            c.priority = UILayoutPriority(999)
             constraints += [
                 bannerView.topAnchor.constraint(equalTo: guide.topAnchor),
-                self.plugin.webView.topAnchor.constraint(equalTo: bannerView.bottomAnchor),
-                self.plugin.webView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
+                c,
+                mainView.topAnchor.constraint(equalTo: bannerView.bottomAnchor)
             ]
         } else {
+            let c = mainView.topAnchor.constraint(equalTo: rootView.topAnchor)
+            c.priority = UILayoutPriority(999)
             constraints += [
                 bannerView.bottomAnchor.constraint(equalTo: guide.bottomAnchor),
-                self.plugin.webView.topAnchor.constraint(equalTo: view.topAnchor),
-                self.plugin.webView.bottomAnchor.constraint(equalTo: bannerView.topAnchor),
+                c,
+                mainView.bottomAnchor.constraint(equalTo: bannerView.topAnchor)
             ]
         }
         NSLayoutConstraint.activate(constraints)
     }
 
     func positionBanner(_ bannerView: UIView) {
-        view.addConstraint(NSLayoutConstraint(item: bannerView,
+        rootView.addConstraint(NSLayoutConstraint(item: bannerView,
                                               attribute: .centerX,
                                               relatedBy: .equal,
-                                              toItem: view,
+                                              toItem: rootView,
                                               attribute: .centerX,
                                               multiplier: 1,
                                               constant: 0))
-        if position == "top" {
-            view.addConstraint(NSLayoutConstraint(item: bannerView,
+        if position == AMSBannerPosition.top {
+            rootView.addConstraint(NSLayoutConstraint(item: bannerView,
                                                   attribute: .top,
                                                   relatedBy: .equal,
-                                                  toItem: plugin.viewController.topLayoutGuide,
+                                                  toItem: rootView.safeAreaLayoutGuide.topAnchor,
                                                   attribute: .top,
                                                   multiplier: 1,
                                                   constant: 0))
         } else {
-            view.addConstraint(NSLayoutConstraint(item: bannerView,
+            rootView.addConstraint(NSLayoutConstraint(item: bannerView,
                                                   attribute: .bottom,
                                                   relatedBy: .equal,
-                                                  toItem: plugin.viewController.bottomLayoutGuide,
+                                                  toItem: rootView.safeAreaLayoutGuide.bottomAnchor,
                                                   attribute: .top,
                                                   multiplier: 1,
                                                   constant: 0))