+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.8](https://github.com/admob-plus/admob-plus/compare/admob-plus@0.15.7...admob-plus@0.15.8) (2020-11-22)
+
+**Note:** Version bump only for package admob-plus
+
+
+
+
+
+## [0.15.7](https://github.com/admob-plus/admob-plus/compare/admob-plus@0.15.6...admob-plus@0.15.7) (2020-10-18)
+
+**Note:** Version bump only for package admob-plus
+
+
+
+
+
+## [0.15.6](https://github.com/admob-plus/admob-plus/compare/admob-plus@0.15.5...admob-plus@0.15.6) (2020-09-12)
+
+**Note:** Version bump only for package admob-plus
+
+
+
+
+
+## [0.15.5](https://github.com/admob-plus/admob-plus/compare/admob-plus@0.15.4...admob-plus@0.15.5) (2020-01-28)
+
+**Note:** Version bump only for package admob-plus
+
+
+
+
+
+## [0.15.4](https://github.com/admob-plus/admob-plus/compare/admob-plus@0.15.3...admob-plus@0.15.4) (2019-09-22)
+
+**Note:** Version bump only for package admob-plus
+
+
+
+
+
 ## [0.15.3](https://github.com/admob-plus/admob-plus/compare/admob-plus@0.15.1...admob-plus@0.15.3) (2019-06-01)