+++ packages/cordova-consent/CHANGELOG.md
@@ -3,6 +3,119 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.18.0](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.17.5...cordova-plugin-consent@0.18.0) (2020-11-22)
+
+
+### Bug Fixes
+
+* **cordova-consent:** fix ios build error ([a73f18f](https://github.com/admob-plus/admob-plus/commit/a73f18fb8797baa3706df3c167b97bb2cdc07a58))
+* **cordova-consent:** update build config ([289c2bd](https://github.com/admob-plus/admob-plus/commit/289c2bdd65cc7cda384a9051b4337e44f206a03b))
+
+
+### Features
+
+* **cordova-consent:** remove usage of cordova-plugin-add-swift-support ([b5ef248](https://github.com/admob-plus/admob-plus/commit/b5ef2489623f11f8b065ef857fd0b3ed959623b4))
+
+
+
+
+
+## [0.17.5](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.17.4...cordova-plugin-consent@0.17.5) (2020-09-12)
+
+
+### Bug Fixes
+
+* **cordova-consent:** copy Android source correctly ([#180](https://github.com/admob-plus/admob-plus/issues/180)) ([bb3e035](https://github.com/admob-plus/admob-plus/commit/bb3e0350952f7b096456c380c1b257ef17c7a8a2))
+
+
+
+
+
+## [0.17.4](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.17.3...cordova-plugin-consent@0.17.4) (2020-09-12)
+
+**Note:** Version bump only for package cordova-plugin-consent
+
+
+
+
+
+## [0.17.3](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.17.2...cordova-plugin-consent@0.17.3) (2020-05-20)
+
+
+### Bug Fixes
+
+* **cordova-consent:** convert consent status to correct string value ([daa6672](https://github.com/admob-plus/admob-plus/commit/daa6672c713d4adf0f0f4c77e1fb22d7bd5a705a))
+
+
+
+
+
+## [0.17.2](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.17.1...cordova-plugin-consent@0.17.2) (2020-05-20)
+
+
+### Bug Fixes
+
+* **cordova-consent:** convert consent status to correct string value ([05aa4d7](https://github.com/admob-plus/admob-plus/commit/05aa4d75da32809f2de10252cb21d5de13b3d203))
+* **cordova-consent:** emit present form events ([2b27c4f](https://github.com/admob-plus/admob-plus/commit/2b27c4fcdd3bd516506a75e688d09a85c4b27e5a))
+
+
+
+
+
+## [0.17.1](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.17.0...cordova-plugin-consent@0.17.1) (2020-05-19)
+
+
+### Bug Fixes
+
+* **cordova-consent:** add ios to supported platforms ([20d1e5e](https://github.com/admob-plus/admob-plus/commit/20d1e5e))
+
+
+
+
+
+# [0.17.0](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.16.1...cordova-plugin-consent@0.17.0) (2020-05-18)
+
+
+### Features
+
+* **cordova-consent:** add iOS support ([6405fd5](https://github.com/admob-plus/admob-plus/commit/6405fd5))
+
+
+
+
+
+## [0.16.1](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.16.0...cordova-plugin-consent@0.16.1) (2020-04-11)
+
+**Note:** Version bump only for package cordova-plugin-consent
+
+
+
+
+
+# [0.16.0](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.15.2...cordova-plugin-consent@0.16.0) (2020-01-28)
+
+
+### Features
+
+* **cordova-consent:** add addTestDevice action ([fe71961](https://github.com/admob-plus/admob-plus/commit/fe71961))
+* **cordova-consent:** add checkConsent API ([1d66bcf](https://github.com/admob-plus/admob-plus/commit/1d66bcf))
+* **cordova-consent:** add ConsentForm ([08b19e3](https://github.com/admob-plus/admob-plus/commit/08b19e3))
+* **cordova-consent:** add isRequestLocationInEeaOrUnknown action ([3179a7d](https://github.com/admob-plus/admob-plus/commit/3179a7d))
+* **cordova-consent:** add setDebugGeography action ([952c928](https://github.com/admob-plus/admob-plus/commit/952c928))
+* **cordova-consent:** prepare for release ([0e9d920](https://github.com/admob-plus/admob-plus/commit/0e9d920))
+
+
+
+
+
+## [0.15.2](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.15.1...cordova-plugin-consent@0.15.2) (2020-01-20)
+
+**Note:** Version bump only for package cordova-plugin-consent
+
+
+
+
+
 ## [0.15.1](https://github.com/admob-plus/admob-plus/compare/cordova-plugin-consent@0.15.0...cordova-plugin-consent@0.15.1) (2019-04-28)
 
 **Note:** Version bump only for package cordova-plugin-consent