+++ packages/ionic-ngx/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.10](https://github.com/admob-plus/admob-plus/compare/@admob-plus/ionic@0.19.9...@admob-plus/ionic@0.19.10) (2020-11-22)
+
+**Note:** Version bump only for package @admob-plus/ionic
+
+
+
+
+
+## [0.19.9](https://github.com/admob-plus/admob-plus/compare/@admob-plus/ionic@0.19.8...@admob-plus/ionic@0.19.9) (2020-09-12)
+
+**Note:** Version bump only for package @admob-plus/ionic
+
+
+
+
+
+## [0.19.8](https://github.com/admob-plus/admob-plus/compare/@admob-plus/ionic@0.19.7...@admob-plus/ionic@0.19.8) (2020-04-11)
+
+**Note:** Version bump only for package @admob-plus/ionic
+
+
+
+
+
+## [0.19.7](https://github.com/admob-plus/admob-plus/compare/@admob-plus/ionic@0.19.6...@admob-plus/ionic@0.19.7) (2020-01-20)
+
+**Note:** Version bump only for package @admob-plus/ionic
+
+
+
+
+
+## [0.19.6](https://github.com/admob-plus/admob-plus/compare/@admob-plus/ionic@0.19.5...@admob-plus/ionic@0.19.6) (2019-09-22)
+
+**Note:** Version bump only for package @admob-plus/ionic
+
+
+
+
+
 ## [0.19.5](https://github.com/admob-plus/admob-plus/compare/@admob-plus/ionic@0.19.4...@admob-plus/ionic@0.19.5) (2019-04-28)
 
 **Note:** Version bump only for package @admob-plus/ionic